+++ core/documentation/source/basic/changelog.rst
@@ -6,6 +6,7 @@ ChangeLog
 
 - Major compatibility on the type() and set_type() functions (content and records classes) on fail
 - Some improvements on build_data(), build_xml() and related() functions of the record class
+- ACL check on the api types() function
 
 
 **v 0.9.6** (2011-11-13)